+++ src/main/scala/org/podval/tools/testing/framework/ZioTest.scala
@@ -3,12 +3,13 @@ package org.podval.tools.testing.framework
 import org.podval.tools.testing.worker.TestTagsFilter
 
 // https://github.com/zio/zio/blob/series/2.x/test-sbt/jvm/src/main/scala/zio/test/sbt/ZTestFramework.scala
-object ZIOTest extends FrameworkDescriptor(
+object ZioTest extends FrameworkDescriptor(
+  // Note: this must be exactly as reported by the framework:
   name = s"${io.AnsiColor.UNDERLINED}ZIO Test${io.AnsiColor.RESET}",
-  displayName = "zio-test",
+  displayName = "ZioTest",
   group = "dev.zio",
   artifact = "zio-test-sbt",
-  versionDefault = "2.0.10",
+  versionDefault = "2.1.14",
   className = "zio.test.sbt.ZTestFramework",
   sharedPackages = List("zio.test.sbt") // TODO more?
 ):