+++ src/main/scala/org/podval/tools/testing/task/TestFrameworkOptions.scala
@@ -1,28 +1,9 @@
 package org.podval.tools.testing.task
 
 import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.testing.junit.JUnitOptions
 import scala.jdk.CollectionConverters.*
 
-// TODO copy the contents of lists to prevent modifications from affecting this
-class TestFrameworkOptions extends org.gradle.api.tasks.testing.TestFrameworkOptions:
-  override def toString: String = s"TestFrameworkOptions(includeTags=$includeTags, excludeTags=$excludeTags)"
-
-  // Note: Gradle Option annotation does not work here to allow setting those from the command line
-  private var includeTags: Set[String] = Set.empty
-  private var excludeTags: Set[String] = Set.empty
-
-  def copyFrom(other: TestFrameworkOptions): Unit =
-    this.includeTags = other.includeTags
-    this.excludeTags = other.excludeTags
-
-  @Input def getIncludeTags: java.util.Set[String] = includeTags.asJava
-  def setIncludeTags(includeTags: java.util.Collection[String]): Unit = this.includeTags = includeTags.asScala.toSet
-  def includeTags(includeTags: String*): TestFrameworkOptions =
-    this.includeTags = this.includeTags ++ includeTags
-    this
-
-  @Input def getExcludeTags: java.util.Set[String] = excludeTags.asJava
-  def setExcludeTags(excludeTags: java.util.Collection[String]): Unit = this.excludeTags = excludeTags.asScala.toSet
-  def excludeTags(excludeTags: String*): TestFrameworkOptions =
-    this.excludeTags = this.excludeTags ++ excludeTags
-    this
+// Note: see org.podval.tools.testing.task.TestTask for the reason of using JUnitOptions.
+class TestFrameworkOptions extends JUnitOptions:
+  override def toString: String = s"TestFrameworkOptions(includeTags=$getIncludeCategories, excludeTags=$getExcludeCategories)"