+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cleanup: minor cleanup
 - chore: minor dependency updates
 - fix: eliminated calls to Task.getProject() during task execution
+- chore: Scala.js dependency update;
+- ZioTest dependency update;
+- cleanup: changed deprecated `org.junit.Assert.assertThat` to `org.hamcrest.MatcherAssert.assertThat`;
+- re-use Gradle's `JUnitOptions` as our TestFrameworkOptions (see the code for the reasons); as a result:
+- `useSbt` block now uses `includeCategories`/`excludeCategories` instead of `includeTags`/`excludeTags`
 
 ## [0.4.14] - 2025-01-07
 - fix: OS detection fails in macOS  https://github.com/opentorah/opentorah/issues/341